161 add data selection functionality to processing set

+++ tests/stakeholder/test_vis_stakeholder.py
@@ -14,7 +14,8 @@ import importlib.resources
 relative_tolerance = 10 ** (-12)
 
 
-def base_test(msv2_name, expected_sum_value):
+def download_and_convert_msv2_to_processing_set(msv2_name):
+    # We can remove this once there is a new release of casacore
     if os.environ["USER"] == "runner":
         casa_data_dir = (importlib.resources.files("casadata") / "__data__").as_posix()
         rc_file = open(os.path.expanduser("~/.casarc"), "a+")  # append mode
@@ -33,11 +34,23 @@ def base_test(msv2_name, expected_sum_value):
         ephemeris_interpolate=True,
         overwrite=True,
     )
+    return ps_name
+
+
+def base_test(file_name, expected_sum_value, is_s3=False):
+
+    if is_s3:
+        ps_name = file_name
+    else:
+        ps_name = download_and_convert_msv2_to_processing_set(file_name)
 
     ps_lazy = read_processing_set(ps_name)
 
     sel_parms = {key: {} for key in ps_lazy.keys()}
     ps = load_processing_set(ps_name, sel_parms=sel_parms)
+
+    ps_lazy_df = ps_lazy.summary()
+    ps_df = ps.summary()
 
     sum = 0.0
     sum_lazy = 0.0
@@ -54,8 +67,9 @@ def base_test(msv2_name, expected_sum_value):
             np.abs(ps_lazy[ms_xds_name][data_name] * ps_lazy[ms_xds_name].WEIGHT)
         )
 
-    os.system("rm -rf " + msv2_name)
-    os.system("rm -rf " + ps_name)
+    if not is_s3:
+        os.system("rm -rf " + file_name)
+        os.system("rm -rf " + ps_name)
 
     assert (
         sum == sum_lazy
@@ -72,42 +86,12 @@ def base_test(msv2_name, expected_sum_value):
             print(f"{xds_name}: {issues}\n")
 
 
-@pytest.mark.parametrize(
-    ("s3_ps_name, expected_sum_value"),
-    [
-        (
-            "s3://viper-test-data/Antennae_North.cal.lsrk.split.v2.vis.zarr",
-            190.0405216217041,
-        )
-    ],
-)
-def test_s3_read_processing_set(s3_ps_name, expected_sum_value):
-    ps_lazy = read_processing_set(s3_ps_name)
-
-    sel_parms = {key: {} for key in ps_lazy.keys()}
-    ps = load_processing_set(s3_ps_name, sel_parms=sel_parms)
-
-    sum = 0.0
-    sum_lazy = 0.0
-
-    for ms_xds_name in ps.keys():
-        if "VISIBILITY" in ps[ms_xds_name]:
-            data_name = "VISIBILITY"
-        else:
-            data_name = "SPECTRUM"
-        sum = sum + np.nansum(
-            np.abs(ps[ms_xds_name][data_name] * ps[ms_xds_name].WEIGHT)
-        )
-        sum_lazy = sum_lazy + np.nansum(
-            np.abs(ps_lazy[ms_xds_name][data_name] * ps_lazy[ms_xds_name].WEIGHT)
-        )
-
-    assert (
-        sum == sum_lazy
-    ), "read_processing_set and load_processing_set VISIBILITY and WEIGHT values differ."
-    assert sum == pytest.approx(
-        expected_sum_value, rel=relative_tolerance
-    ), "VISIBILITY and WEIGHT values have changed."
+def test_s3():
+    base_test(
+        "s3://viper-test-data/Antennae_North.cal.lsrk.split.v3.vis.zarr",
+        190.0405216217041,
+        is_s3=True,
+    )
 
 
 def test_alma():
@@ -146,6 +130,7 @@ def test_single_dish():
     base_test("sdimaging.ms", 5487446.5)
 
 
+# test_s3()
 # test_alma()
 # test_ska_mid()
 # test_lofar()
@@ -153,5 +138,5 @@ def test_single_dish():
 # test_global_vlbi()
 # test_vlba()
 # test_ngeht()
-test_ephemeris()
+# test_ephemeris()
 # test_single_dish()
